requirements.txt
streamlit
torch
torchvision
pillow
scikit-learn
matplotlib
gdown
